requirements.txt
anyio==3.7.1
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
numpy==1.21.6
prettytable==3.7.0
python-dotenv==0.21.1
python-telegram-bot==20.3
scipy==1.7.3
sniffio==1.3.0
typing_extensions==4.7.1
wcwidth==0.2.9
zipp==3.15.0
